pyproject.toml
[project]
name = "protected-fishing-areas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow==2.9.3",
    "apache-airflow-providers-postgres>=6.1.3",
    "dbt-postgres>=1.10.0",
    "fastapi>=0.128.0",
    "fiona>=1.10.1",
    "geoalchemy2>=0.18.1",
    "geopandas>=1.1.2",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "pyproj>=3.7.2",
    "shapely>=2.1.2",
    "sqlalchemy>=1.4.54",
    "uvicorn>=0.40.0",
]
